worker/requirements.txt
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.34.69
watchdog
redis==5.0.1
